✨ feat: initialize service and database enviroment
