+++ .env.development
@@ -6,5 +6,5 @@ DATABASE_URL="postgresql://lawdlink_user:password@localhost:5432/lawdlink_db?sch
 
 SERVICE_API_VERSION=v1
 SERVICE_API_GATEWAY_PORT=80
-SERVICE_API_AUTH_PORT=3001
-SERVICE_API_USER_PORT=3002
+SERVICE_API_USER_PORT=4001
+SERVICE_API_AUTH_PORT=4002
