+++ .eslintrc
@@ -9,6 +9,7 @@
     "plugin:@typescript-eslint/recommended"
   ],
   "rules": { 
-    "no-console": 1
+    "no-console": 1,
+    "@typescript-eslint/no-namespace": 0
   }
 }
