+++ .vscode/settings.json
@@ -1,0 +1,14 @@
+{
+    "rest-client.environmentVariables": {
+        "$shared": {
+            "version": "v1",
+            "prodToken": "foo",
+            "nonProdToken": "bar"
+        },
+        "local": {
+            "version": "{{$shared version}}",
+            "host": "http://localhost:80",
+            "secretKey": "devSecret"
+        }
+    }
+}
