+++ api.rest
@@ -1,0 +1,11 @@
+### CHECK HEALTH
+GET {{host}}
+Content-Type: application/json
+
+### V1 HEALTH 
+GET {{host}}/v1
+Content-Type: application/json
+
+### V1 USER
+GET {{host}}/v1/user/
+Content-Type: application/json
