+++ package.json
@@ -2,11 +2,13 @@
   "name": "law-d-link",
   "private": true,
   "scripts": {
-    "seed": "bun --bun run prisma/seeds/index.ts",
-    "dev": "bun --bun concurrently -c -n api-gateway,api-auth,api-user \"cd services/api-gateway && bun run dev\" \"cd services/api-auth && bun run dev\" \"cd services/api-user && bun run dev\"",
-    "lint": "bun --bun concurrently -c -n api-gateway,api-auth,api-user \"cd services/api-gateway && bun run lint\" \"cd services/api-auth && bun run lint\" \"cd services/api-user && bun run lint\"",
-    "build:bin": "bun --bun concurrently -c -n api-gateway,api-auth,api-user \"cd services/api-gateway && bun run build:bin\" \"cd services/api-auth && bun run build:bin\" \"cd services/api-user && bun run build:bin\"",
-    "build:openapi": "mkdir -p dist/openapi-doc && ./node_modules/.bin/redocly build-docs docs/OpenAPI/link-d-law.json --config infrastructures/redoc/redocly.yaml -o dist/openapi-doc/index.html"
+    "dev": "bun --bun concurrently -c -n api-gateway,api-auth,api-main \"cd services/api-gateway && bun run dev\" \"cd services/api-auth && bun run dev\" \"cd services/api-main && bun run dev\"",
+    "lint": "bun --bun concurrently -c -n api-gateway,api-auth,api-main \"cd services/api-gateway && bun run lint\" \"cd services/api-auth && bun run lint\" \"cd services/api-main && bun run lint\"",
+    "build:bin": "bun --bun concurrently -c -n api-gateway,api-auth,api-main \"cd services/api-gateway && bun run build:bin\" \"cd services/api-auth && bun run build:bin\" \"cd services/api-main && bun run build:bin\"",
+    "build:openapi": "mkdir -p dist/openapi-doc && ./node_modules/.bin/redocly build-docs docs/OpenAPI/link-d-law.json --config infrastructures/redoc/redocly.yaml -o dist/openapi-doc/index.html",
+    "db:seed": "bun --bun run prisma/seeds/index.ts",
+    "db:prepare": "bunx prisma format && bunx prisma generate",
+    "db:migrate:dev": "dotenv -e .env.development -- bunx prisma migrate dev && bunx db:prepare"
   },
   "devDependencies": {
     "@redocly/cli": "^1.4.1",
@@ -14,6 +16,7 @@
     "@typescript-eslint/parser": "^6.10.0",
     "bun-types": "latest",
     "concurrently": "^8.2.2",
+    "dotenv-cli": "^7.3.0",
     "eslint": "^8.53.0"
   },
   "peerDependencies": {
