+++ services/api-main/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@law-d-link/api-user",
+  "name": "@law-d-link/api-main",
   "module": "src/index.ts",
   "scripts": {
     "start": "bun --bun run src/index.ts",
     "dev": "bun --bun --watch run src/index.ts",
     "dev:nodemon": "nodemon --ext ts  --watch 'src/**/*' --exec 'bun run start' --quiet",
     "lint": "eslint --config ../../.eslintrc --ext .ts src",
-    "build:bin": "bun build ./src/index.ts --compile --outfile ../../bin/service-api-user"
+    "build:bin": "bun build ./src/index.ts --compile --outfile ../../bin/service-api-main"
   },
   "dependencies": {
     "express": "^4.18.2"
