+++ services/api-main/tsconfig.json
